+++ HISTORY.md
@@ -8,6 +8,7 @@
 * Fix broken noop on `acl clear`.
 * Fix `derivedmetric set` command.
 * Fix various `alert` commands when using `--noop`.
+* Better handle bad `search` input.
 
 ## 4.2.1 (26/07/2019)
 * Remove tagging subcommands from `apitoken` command's usage info,